refactor: abstract query management
